+++ resolwe_bio/process/models.py
@@ -1,4 +1,7 @@
 """Additional models used by ProcessBio."""
+from typing import Any
+
+from resolwe.process.communicator import communicator
 from resolwe.process.models import Model
 
 
@@ -18,6 +21,28 @@ class Feature(Model):
         "description",
         "aliases",
     ]
+
+    def get_features(
+        self,
+        source,
+        species,
+        feature_ids,
+        required_fields=(
+            "source",
+            "feature_id",
+            "species",
+            "type",
+            "sub_type",
+            "name",
+            "full_name",
+            "description",
+            "aliases",
+        ),
+    ) -> list:
+        """Get features from the database."""
+        return communicator.filter_features(
+            (source, species, feature_ids, required_fields)
+        )
 
 
 class Mapping(Model):