+++ docs/CHANGELOG.rst
@@ -15,6 +15,8 @@ Added
 -----
 - Add command ``filter_features`` to the listener that caches the returned
   ``Feature`` objects to speed up the queries
+- Add command ``get_features`` to the ``Feature`` class 
+  
 
 Changed
 -------